[mediasdk][codec] Fix Coverity static analysis issues
Fix signed integer overflow (CID 3515584, 3544062) in MPEG2 BRC
code where int32_t arithmetic could overflow before assignment to
uint32_t destinations.

- umc_mpeg2_brc.cpp (CheckHRDParams): Cast mParams.HRDBufferSizeBytes
  to uint32_t before multiplying by 8 to avoid signed overflow.
  mParams.HRDBufferSizeBytes can be up to 536866816 and multiplying
  by 8 as int32_t exceeds INT32_MAX (undefined behavior in C++).

- umc_video_brc.cpp (CheckCorrectParams_MPEG2): Cast bufferSizeBytes
  to uint32_t before left-shifting by 3 to avoid signed overflow.
  bufferSizeBytes is capped at 536870911 (MFX_MAX_32U >> 3) and
  shifting as int32_t exceeds INT32_MAX (undefined behavior in C++).

CIDs 3583108 (VVC InitScanOrderTables y-- underflow) and 3583193
(AV1 order_hint_bits_minus1 overflow) are false positives:
- CID 3583108: The if (y == 0 || ...) condition at line 98 prevents
  the else branch (y--) from executing when y is 0.
- CID 3583193: av1_set_frame_refs() is only called when
  sh.enable_order_hint is true, ensuring order_hint_bits_minus1
  is never -1 at that call site.

+++ _studio/shared/umc/codec/brc/src/umc_mpeg2_brc.cpp
@@ -240,7 +240,7 @@ Status MPEG2BRC::CheckHRDParams()
     }
   }
   mParams.HRDBufferSizeBytes = mHRD.bufSize / 8;
-  mHRD.bufSize = mParams.HRDBufferSizeBytes * 8;
+  mHRD.bufSize = (uint32_t)mParams.HRDBufferSizeBytes * 8;
   mParams.HRDInitialDelayBytes = (int32_t)(mHRD.bufFullness / 8);
   mHRD.bufFullness = mParams.HRDInitialDelayBytes * 8;
